De-duplicate the ECDSA tests
Rather than duplicating all the bosh client tests, we can make do with
just one. When a server is configured with an ECDSA public key, we
check that the handshake method used is the ecdsa-sha2-nistp521 one.

[#182248334]
